refactor: update logic for deviceCode
